mistral/agents/agents_api/food_diet_companion/requirements.txt
--index-url https://2kUrwd-RY03seHskbjvagCk3bmdnwsM3U@pypi.fury.io/mistralai/
--extra-index-url https://pypi.org/simple

chainlit==2.3.0
python-dotenv==1.1.0
mistralai-private==1.3.8rc71
pydantic==2.10.5
loguru==0.7.3
tavily-python==0.5.0
typing-extensions==4.12.2
anyio==4.8.0
annotated-types==0.7.0 